LayoutTests/webanimations/Document-expected.txt

Web Animation API: Document interface extension test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.timeline is not null
PASS document.timeline is document.timeline
PASS document.timeline is not iframe.contentDocument.timeline
PASS document.getAnimations().length is 0
PASS target.getAnimations().length is 0
PASS iframe.contentDocument.getAnimations().length is 0
PASS animation is not null
PASS document.getAnimations().length is 1
PASS target.getAnimations().length is 1
PASS iframe.contentDocument.getAnimations().length is 0

PASS successfullyParsed is true

TEST COMPLETE
